requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.14
